tests: package rename net_testutils to testutils
